Auto-Coder: Address issue #461
