+++ src/auto_coder/automation_engine.py
@@ -16,7 +16,8 @@ from . import fix_to_pass_tests_runner as fix_to_pass_tests_runner_module
 from .automation_config import AutomationConfig, Candidate, CandidateProcessingResult, ProcessResult
 from .fix_to_pass_tests_runner import fix_to_pass_tests
 from .gh_logger import get_gh_logger
-from .git_utils import git_commit_with_retry, git_push
+from .git_branch import git_commit_with_retry
+from .git_commit import git_push
 from .issue_processor import create_feature_issues
 from .logger_config import get_logger
 from .pr_processor import _create_pr_analysis_prompt as _engine_pr_prompt
