+++ src/auto_coder/fix_to_pass_tests_runner.py
@@ -11,8 +11,9 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Any, Dict, Optional, Tuple
 
 from .automation_config import AutomationConfig
+from .git_branch import git_commit_with_retry
+from .git_commit import git_push, save_commit_failure_history
 from .git_info import get_commit_log
-from .git_utils import git_commit_with_retry, git_push, save_commit_failure_history
 from .logger_config import get_logger, log_calls
 from .progress_footer import ProgressStage
 from .prompt_loader import render_prompt
