+++ tests/test_git_utils.py
@@ -13,7 +13,12 @@ from src.auto_coder.git_branch import (
     branch_exists,
     extract_number_from_branch,
     git_checkout_branch,
+    git_commit_with_retry,
     validate_branch_name,
+)
+from src.auto_coder.git_commit import (
+    git_push,
+    save_commit_failure_history,
 )
 from src.auto_coder.git_info import (
     get_commit_log,
@@ -26,10 +31,7 @@ from src.auto_coder.git_utils import (
     branch_context,
     get_all_branches,
     get_branches_by_pattern,
-    git_commit_with_retry,
-    git_push,
     migrate_pr_branches,
-    save_commit_failure_history,
 )
 from src.auto_coder.utils import CommandResult
 
@@ -40,7 +42,7 @@ class TestGitCommitWithRetry:
 
     def test_successful_commit(self):
         """Test successful commit without retry."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(success=True, stdout="", stderr="", returncode=0)
@@ -54,7 +56,7 @@ class TestGitCommitWithRetry:
 
     def test_commit_with_dprint_error_and_retry(self):
         """Test commit with dprint formatting error triggers retry."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
 
@@ -87,7 +89,7 @@ class TestGitCommitWithRetry:
 
     def test_commit_with_dprint_error_fmt_fails(self):
         """Test commit with dprint error but formatter fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
 
@@ -122,7 +124,7 @@ class TestGitCommitWithRetry:
 
     def test_commit_with_unknown_error_llm_fallback_and_retry_success(self):
         """Unknown commit error triggers LLM fallback and retry commit succeeds."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor, patch("src.auto_coder.git_utils.try_llm_commit_push") as mock_llm:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor, patch("src.auto_coder.git_branch.try_llm_commit_push") as mock_llm:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_llm.return_value = True
@@ -141,7 +143,7 @@ class TestGitCommitWithRetry:
 
     def test_commit_with_unknown_error_llm_fallback_nothing_to_commit(self):
         """Unknown commit error triggers LLM fallback; treat as success when nothing left to commit."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor, patch("src.auto_coder.git_utils.try_llm_commit_push") as mock_llm:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor, patch("src.auto_coder.git_branch.try_llm_commit_push") as mock_llm:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_llm.return_value = True
@@ -160,7 +162,7 @@ class TestGitCommitWithRetry:
 
     def test_commit_with_cwd(self):
         """Test commit with custom working directory."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_branch.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(success=True, stdout="", stderr="", returncode=0)
@@ -178,7 +180,7 @@ class TestGitPush:
 
     def test_successful_push(self):
         """Test successful push without branch specified."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -203,7 +205,7 @@ class TestGitPush:
 
     def test_push_with_branch(self):
         """Test push with specific branch."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -220,7 +222,7 @@ class TestGitPush:
 
     def test_push_with_custom_remote(self):
         """Test push with custom remote."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -237,7 +239,7 @@ class TestGitPush:
 
     def test_push_failure(self):
         """Test push failure."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -264,7 +266,7 @@ class TestGitPush:
 
     def test_push_with_cwd(self):
         """Test push with custom working directory."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -291,7 +293,7 @@ class TestGitPush:
 
     def test_push_no_upstream_auto_retry(self):
         """Test push automatically retries with --set-upstream when upstream is not set."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -331,7 +333,7 @@ class TestGitPush:
 
     def test_push_no_upstream_with_branch_specified(self):
         """Test push with branch specified automatically retries with --set-upstream."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -363,7 +365,7 @@ class TestGitPush:
 
     def test_push_other_error_no_retry(self):
         """Test push does not retry for errors other than missing upstream."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -391,7 +393,7 @@ class TestGitPush:
 
     def test_push_with_dprint_error_and_retry(self):
         """Test push with dprint formatting error triggers retry without commit message."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -429,7 +431,7 @@ class TestGitPush:
 
     def test_push_with_dprint_error_and_commit_message(self):
         """Test push with dprint formatting error and commit message triggers re-commit."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -470,7 +472,7 @@ class TestGitPush:
 
     def test_push_with_dprint_error_fmt_fails(self):
         """Test push with dprint error but formatter fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -505,7 +507,7 @@ class TestGitPush:
 
     def test_push_with_dprint_error_commit_amend_fails(self):
         """Test push with dprint error when commit amend fails, falls back to regular commit."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -552,7 +554,7 @@ class TestGitPush:
 
     def test_push_with_dprint_error_and_upstream_retry(self):
         """Test push with dprint error and then upstream error triggers both retries."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor_utils, patch("src.auto_coder.git_info.CommandExecutor") as mock_executor_info:
             mock_cmd = MagicMock()
             mock_executor_utils.return_value = mock_cmd
             mock_executor_info.return_value = mock_cmd
@@ -735,7 +737,7 @@ class TestGitCheckoutBranch:
 
     def test_successful_checkout_existing_branch(self):
         """Test successful checkout of an existing branch."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -778,7 +780,7 @@ class TestGitCheckoutBranch:
 
     def test_successful_checkout_create_new_branch(self):
         """Test successful checkout with creating a new branch from origin/main."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -854,7 +856,7 @@ class TestGitCheckoutBranch:
 
     def test_successful_checkout_create_from_base_branch(self):
         """Test successful checkout with creating a new branch from base branch with base ref included."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -921,7 +923,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_failure(self):
         """Test checkout failure."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -945,7 +947,7 @@ class TestGitCheckoutBranch:
 
     def test_create_new_branch_requires_base_branch(self):
         """Creating a new branch without base_branch should raise ValueError."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # First call for branch list (branch doesn't exist)
@@ -956,7 +958,7 @@ class TestGitCheckoutBranch:
 
     def test_create_new_branch_fallback_to_local_base(self):
         """When origin/<base_branch> is missing, fall back to local <base_branch>."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1005,7 +1007,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_success_but_verification_fails(self):
         """Test checkout succeeds but verification command fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1035,7 +1037,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_success_but_branch_mismatch(self):
         """Test checkout succeeds but current branch doesn't match expected."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1062,7 +1064,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_with_cwd(self):
         """Test checkout with custom working directory."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1089,7 +1091,7 @@ class TestGitCheckoutBranch:
 
     def test_create_new_branch_push_failure(self):
         """Test creating a new branch when push fails (should still succeed)."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1127,7 +1129,7 @@ class TestGitCheckoutBranch:
 
     def test_create_new_branch_without_publish(self):
         """Test creating a new branch without publishing to remote."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1158,7 +1160,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_with_uncommitted_changes_auto_commit(self):
         """Test checkout with uncommitted changes automatically commits them."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1200,7 +1202,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_with_uncommitted_changes_error_retry(self):
         """Test checkout fails with uncommitted changes error, then retries after commit."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1255,7 +1257,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_with_uncommitted_changes_commit_fails(self):
         """Test checkout with uncommitted changes when commit fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.side_effect = [
@@ -1288,7 +1290,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_rejects_invalid_pr_branch_name_when_creating(self):
         """Test that creating new branch with pr-<number> pattern is rejected."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # Mock that branch doesn't exist
@@ -1304,7 +1306,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_accepts_existing_pr_branch(self):
         """Test that checking out existing pr-<number> branch is allowed when not creating a new branch."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # When create_new=False, no validation is performed; allow pr-<number>
@@ -1324,7 +1326,7 @@ class TestGitCheckoutBranch:
 
     def test_checkout_rejects_invalid_pr_branch_name_case_insensitive(self):
         """Test that creating pr-<number> pattern is rejected regardless of case."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # Mock that branch doesn't exist
@@ -1602,7 +1604,7 @@ class TestGetAllBranches:
 
     def test_get_all_branches_local(self):
         """Test getting all local branches."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1619,7 +1621,7 @@ class TestGetAllBranches:
 
     def test_get_all_branches_remote(self):
         """Test getting all remote branches."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1636,7 +1638,7 @@ class TestGetAllBranches:
 
     def test_get_all_branches_empty(self):
         """Test getting branches when none exist."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(success=True, stdout="", stderr="", returncode=0)
@@ -1647,7 +1649,7 @@ class TestGetAllBranches:
 
     def test_get_all_branches_with_cwd(self):
         """Test getting branches with custom working directory."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1664,7 +1666,7 @@ class TestGetAllBranches:
 
     def test_get_all_branches_failure(self):
         """Test getting branches when git command fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1684,7 +1686,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_with_wildcard(self):
         """Test getting branches matching a pattern with wildcard."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # get_all_branches is called internally
@@ -1702,7 +1704,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_exact_match(self):
         """Test getting branches with exact match pattern."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1719,7 +1721,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_case_insensitive(self):
         """Test case-insensitive pattern matching."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1736,7 +1738,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_no_matches(self):
         """Test pattern matching when no branches match."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1753,7 +1755,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_with_remote_prefix(self):
         """Test pattern matching with remote branch prefix."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             mock_cmd.run_command.return_value = CommandResult(
@@ -1770,7 +1772,7 @@ class TestGetBranchesByPattern:
 
     def test_get_branches_by_pattern_local_only(self):
         """Test pattern matching in local branches only."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # Mock for local branches (no -r flag)
@@ -1795,7 +1797,7 @@ class TestMigratePrBranches:
 
     def test_migrate_pr_branches_no_pr_branches(self):
         """Test migration when no pr-<number> branches exist."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # git branch -r (get_all_branches) - no pr-* branches
@@ -1954,7 +1956,7 @@ class TestMigratePrBranches:
 
     def test_migrate_pr_branches_extraction_failure(self):
         """Test migration when number extraction fails for a branch."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             mock_cmd = MagicMock()
             mock_executor.return_value = mock_cmd
             # Branch with invalid pattern (no number)
@@ -2090,7 +2092,7 @@ class TestBranchContext:
 
     def test_branch_context_successful_switch_and_return(self):
         """Test successful branch switch and return using context manager."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2120,7 +2122,7 @@ class TestBranchContext:
 
     def test_branch_context_with_exception(self):
         """Test that context manager returns to original branch even when exception occurs."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2145,7 +2147,7 @@ class TestBranchContext:
 
     def test_branch_context_create_new_branch(self):
         """Test context manager with create_new parameter."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2173,7 +2175,7 @@ class TestBranchContext:
 
     def test_branch_context_with_custom_cwd(self):
         """Test context manager with custom working directory."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2195,7 +2197,7 @@ class TestBranchContext:
 
     def test_branch_context_already_on_target_branch(self):
         """Test context manager when already on the target branch."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2216,7 +2218,7 @@ class TestBranchContext:
 
     def test_branch_context_switch_failure(self):
         """Test that RuntimeError is raised when switch_to_branch fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2255,7 +2257,7 @@ class TestBranchContext:
 
     def test_branch_context_not_git_repository_on_exit(self):
         """Test that context manager handles not being in a git repo during cleanup."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2277,7 +2279,7 @@ class TestBranchContext:
 
     def test_branch_context_return_switch_failure(self):
         """Test that return switch failure is logged but doesn't raise."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2302,7 +2304,7 @@ class TestBranchContext:
 
     def test_branch_context_with_pull_after_switch(self):
         """Test that pull_after_switch is always True for both entry and exit."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2323,7 +2325,7 @@ class TestBranchContext:
 
     def test_branch_context_checks_unpushed_commits_by_default(self):
         """Test that branch_context checks and pushes unpushed commits by default."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2352,7 +2354,7 @@ class TestBranchContext:
 
     def test_branch_context_skips_unpushed_commits_when_disabled(self):
         """Test that branch_context skips unpushed commit check when check_unpushed=False."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2373,7 +2375,7 @@ class TestBranchContext:
 
     def test_branch_context_with_custom_remote(self):
         """Test that branch_context uses custom remote for unpushed commit check."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
@@ -2401,7 +2403,7 @@ class TestBranchContext:
 
     def test_branch_context_handles_ensure_pushed_failure(self):
         """Test that branch_context continues even if ensure_pushed fails."""
-        with patch("src.auto_coder.git_utils.CommandExecutor") as mock_executor:
+        with patch("src.auto_coder.git_commit.CommandExecutor") as mock_executor:
             with patch("src.auto_coder.git_utils.get_current_branch") as mock_get_branch:
                 with patch("src.auto_coder.git_utils.switch_to_branch") as mock_switch:
                     with patch("src.auto_coder.git_utils.is_git_repository") as mock_is_repo:
